+++ ui.js
@@ -13,5 +13,6 @@ export { renderReview } from "./views/review.js";
 export { renderYearReview } from "./views/yearreview.js";
 export { renderQuests } from "./views/quests.js";
 export { renderLeaderboard } from "./views/leaderboard.js";
+export { renderProfile } from "./views/profile.js";
 export { getLumiTip } from "./views/assistant.js";
 export { openDialog } from "./views/helpers.js";